responsable + solucion todo lo de servicio

+++ src/main/java/es/uji/ei102720mgph/SANA/controller/MunicipalManagerController.java
@@ -7,6 +7,7 @@ import es.uji.ei102720mgph.SANA.dao.SanaUserDao;
 import es.uji.ei102720mgph.SANA.model.Email;
 import es.uji.ei102720mgph.SANA.model.MunicipalManager;
 import es.uji.ei102720mgph.SANA.model.Municipality;
+import es.uji.ei102720mgph.SANA.model.UserLogin;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -16,6 +17,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
+import javax.servlet.http.HttpSession;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -47,7 +49,12 @@ public class MunicipalManagerController {
     }
 
     @RequestMapping("/list")
-    public String listMunicipalManager(Model model) {
+    public String listMunicipalManager(Model model, HttpSession session) {
+        if(session.getAttribute("environmentalManager") ==  null) {
+            model.addAttribute("userLogin", new UserLogin() {});
+            session.setAttribute("nextUrl", "/municipalManager/list");
+            return "redirect:/inicio/login";
+        }
         model.addAttribute("municipalManagers", municipalManagerDao.getMunicipalManagers());
         return "municipalManager/list";
     }
@@ -63,7 +70,12 @@ public class MunicipalManagerController {
     }
 
     @RequestMapping(value="/add")
-    public String addMunicipalManager(Model model) {
+    public String addMunicipalManager(Model model, HttpSession session) {
+        if(session.getAttribute("environmentalManager") ==  null) {
+            model.addAttribute("userLogin", new UserLogin() {});
+            session.setAttribute("nextUrl", "/municipalManager/add");
+            return "redirect:/inicio/login";
+        }
         model.addAttribute("municipalManager", new MunicipalManager());
         return "municipalManager/add";
     }
@@ -99,7 +111,12 @@ public class MunicipalManagerController {
     }
 
     @RequestMapping(value="/update/{email}", method = RequestMethod.GET)
-    public String editMunicipalManager(Model model, @PathVariable String email) {
+    public String editMunicipalManager(Model model, @PathVariable String email, HttpSession session) {
+        if(session.getAttribute("environmentalManager") ==  null) {
+            model.addAttribute("userLogin", new UserLogin() {});
+            session.setAttribute("nextUrl", "/municipalManager/update/" + email);
+            return "redirect:/inicio/login";
+        }
         model.addAttribute("municipalManager", municipalManagerDao.getMunicipalManager(email));
         return "municipalManager/update";
     }
@@ -117,7 +134,12 @@ public class MunicipalManagerController {
     }
 
     @RequestMapping(value="/darDeBaja/{email}", method = RequestMethod.GET)
-    public String darDeBajaMunicipalManager(@PathVariable String email) {
+    public String darDeBajaMunicipalManager(Model model, @PathVariable String email,HttpSession session) {
+        if(session.getAttribute("environmentalManager") ==  null) {
+            model.addAttribute("userLogin", new UserLogin() {});
+            session.setAttribute("nextUrl", "/municipalManager/darDeBaja/" + email);
+            return "redirect:/inicio/login";
+        }
         MunicipalManager municipalManager = municipalManagerDao.getMunicipalManager(email);
         municipalManager.setLeavingDate(LocalDate.now());
         municipalManagerDao.updateMunicipalManager(municipalManager);
@@ -125,7 +147,12 @@ public class MunicipalManagerController {
     }
 
     @RequestMapping(value="/delete/{email}")
-    public String processDelete(@PathVariable String email) {
+    public String processDelete(Model model, @PathVariable String email, HttpSession session) {
+        if(session.getAttribute("environmentalManager") ==  null) {
+            model.addAttribute("userLogin", new UserLogin() {});
+            session.setAttribute("nextUrl", "/municipalManager/delete/" + email);
+            return "redirect:/inicio/login";
+        }
         municipalManagerDao.deleteMunicipalManager(email);
         sanaUserDao.deleteSanaUser(email);
         return "redirect:../list";
