+++ src/main/java/es/uji/ei102720mgph/SANA/controller/MunicipalityController.java
@@ -5,6 +5,7 @@ import es.uji.ei102720mgph.SANA.dao.MunicipalityDao;
 import es.uji.ei102720mgph.SANA.dao.PostalCodeMunicipalityDao;
 import es.uji.ei102720mgph.SANA.model.Municipality;
 import es.uji.ei102720mgph.SANA.model.PostalCodeMunicipality;
+import es.uji.ei102720mgph.SANA.model.UserLogin;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -13,6 +14,8 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+
+import javax.servlet.http.HttpSession;
 
 @Controller
 @RequestMapping("/municipality")
@@ -38,13 +41,23 @@ public class MunicipalityController {
     }
 
     @RequestMapping("/list")
-    public String listMunicipalities(Model model) {
+    public String listMunicipalities(Model model, HttpSession session) {
+        if(session.getAttribute("environmentalManager") ==  null) {
+            model.addAttribute("userLogin", new UserLogin() {});
+            session.setAttribute("nextUrl", "/municipality/list");
+            return "redirect:/inicio/login";
+        }
         model.addAttribute("municipalities", municipalityDao.getMunicipalities());
         return "municipality/list";
     }
 
     @RequestMapping(value="/get/{name}")
-    public String getMunicipality(Model model, @PathVariable("name") String name){
+    public String getMunicipality(Model model, @PathVariable("name") String name, HttpSession session){
+        if(session.getAttribute("environmentalManager") ==  null) {
+            model.addAttribute("userLogin", new UserLogin() {});
+            session.setAttribute("nextUrl", "/municipality/get/" + name);
+            return "redirect:/inicio/login";
+        }
         model.addAttribute("municipality", municipalityDao.getMunicipality(name));
         model.addAttribute("municipalManagers", municipalManagerDao.getManagersOfMunicipality(name));
         model.addAttribute("postalCodes", postalCodeMunicipalityDao.getPostalCodeOfMuni(name));
@@ -55,7 +68,12 @@ public class MunicipalityController {
     }
 
     @RequestMapping(value="/add")
-    public String addMunicipality(Model model) {
+    public String addMunicipality(Model model, HttpSession session) {
+        if(session.getAttribute("environmentalManager") ==  null) {
+            model.addAttribute("userLogin", new UserLogin() {});
+            session.setAttribute("nextUrl", "/municipality/add");
+            return "redirect:/inicio/login";
+        }
         model.addAttribute("municipality", new Municipality());
         return "municipality/add";
     }
@@ -73,7 +91,12 @@ public class MunicipalityController {
     }
 
     @RequestMapping(value="/update/{name}", method = RequestMethod.GET)
-    public String editMunicipality(Model model, @PathVariable String name) {
+    public String editMunicipality(Model model, @PathVariable String name, HttpSession session) {
+        if(session.getAttribute("environmentalManager") ==  null) {
+            model.addAttribute("userLogin", new UserLogin() {});
+            session.setAttribute("nextUrl", "/municipality/update/" + name);
+            return "redirect:/inicio/login";
+        }
         model.addAttribute("municipality", municipalityDao.getMunicipality(name));
         return "municipality/update";
     }
@@ -91,7 +114,12 @@ public class MunicipalityController {
     }
 
     @RequestMapping(value="/delete/{name}")
-    public String processDelete(@PathVariable String name) {
+    public String processDelete(Model model, @PathVariable String name, HttpSession session) {
+        if(session.getAttribute("environmentalManager") ==  null) {
+            model.addAttribute("userLogin", new UserLogin() {});
+            session.setAttribute("nextUrl", "/municipality/delete/" + name);
+            return "redirect:/inicio/login";
+        }
         municipalityDao.deleteMunicipality(name);
         return "redirect:../list";
     }
