+++ src/main/java/es/uji/ei102720mgph/SANA/controller/PostalCodeMunicipalityController.java
@@ -2,6 +2,7 @@ package es.uji.ei102720mgph.SANA.controller;
 
 import es.uji.ei102720mgph.SANA.dao.PostalCodeMunicipalityDao;
 import es.uji.ei102720mgph.SANA.model.PostalCodeMunicipality;
+import es.uji.ei102720mgph.SANA.model.UserLogin;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -10,6 +11,8 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+
+import javax.servlet.http.HttpSession;
 
 @Controller
 @RequestMapping("/postalCode")
@@ -23,7 +26,12 @@ public class PostalCodeMunicipalityController {
     }
 
     @RequestMapping(value="/add")
-    public String addPostalCode(Model model, @PathVariable String municipality) {
+    public String addPostalCode(Model model, @PathVariable String municipality, HttpSession session) {
+        if(session.getAttribute("environmentalManager") ==  null) {
+            model.addAttribute("userLogin", new UserLogin() {});
+            session.setAttribute("nextUrl", "/postalCode/add");
+            return "redirect:/inicio/login";
+        }
         model.addAttribute("postalCode", new PostalCodeMunicipality());
         return "postalCode/add";
     }
@@ -41,7 +49,12 @@ public class PostalCodeMunicipalityController {
     }
 
     @RequestMapping(value="/delete/{municipality}/{postalCode}")
-    public String processDelete(@PathVariable String municipality,@PathVariable String postalCode) {
+    public String processDelete(Model model, @PathVariable String municipality,@PathVariable String postalCode, HttpSession session) {
+        if(session.getAttribute("environmentalManager") ==  null) {
+            model.addAttribute("userLogin", new UserLogin() {});
+            session.setAttribute("nextUrl", "/postalCode/delete/" + municipality + "/" + postalCode);
+            return "redirect:/inicio/login";
+        }
         pcD.deletePostalCode(municipality, postalCode);
         return "redirect:/municipality/get/" + municipality;
     }
