+++ src/main/java/es/uji/ei102720mgph/SANA/controller/ServiceController.java
@@ -5,6 +5,7 @@ import es.uji.ei102720mgph.SANA.enums.Temporality;
 import es.uji.ei102720mgph.SANA.enums.TypeOfAccess;
 import es.uji.ei102720mgph.SANA.model.Municipality;
 import es.uji.ei102720mgph.SANA.model.Service;
+import es.uji.ei102720mgph.SANA.model.UserLogin;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -14,6 +15,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
+import javax.servlet.http.HttpSession;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -30,14 +32,24 @@ public class ServiceController {
 
     // Operació llistar
     @RequestMapping("/list")
-    public String listServices(Model model) {
+    public String listServices(Model model, HttpSession session) {
+        if(session.getAttribute("environmentalManager") ==  null) {
+            model.addAttribute("userLogin", new UserLogin() {});
+            session.setAttribute("nextUrl", "/service/list");
+            return "redirect:/inicio/login";
+        }
         model.addAttribute("services", serviceDao.getServices());
         return "service/list";
     }
 
     // Operació crear
     @RequestMapping(value="/add")
-    public String addService(Model model) {
+    public String addService(Model model, HttpSession session) {
+        if(session.getAttribute("environmentalManager") ==  null) {
+            model.addAttribute("userLogin", new UserLogin() {});
+            session.setAttribute("nextUrl", "/service/add");
+            return "redirect:/inicio/login";
+        }
         model.addAttribute("service", new Service());
         return "service/add";
     }
@@ -57,7 +69,12 @@ public class ServiceController {
 
     // Operació actualitzar
     @RequestMapping(value="/update/{nameOfService}", method = RequestMethod.GET)
-    public String editService(Model model, @PathVariable String nameOfService) {
+    public String editService(Model model, @PathVariable String nameOfService, HttpSession session) {
+        if(session.getAttribute("environmentalManager") ==  null) {
+            model.addAttribute("userLogin", new UserLogin() {});
+            session.setAttribute("nextUrl", "/service/update/" + nameOfService);
+            return "redirect:/inicio/login";
+        }
         model.addAttribute("service", serviceDao.getService(nameOfService));
         return "service/update";
     }
@@ -77,7 +94,12 @@ public class ServiceController {
 
     // Operació esborrar
     @RequestMapping(value="/delete/{nameOfService}")
-    public String processDelete(@PathVariable String nameOfService) {
+    public String processDelete(Model model, @PathVariable String nameOfService, HttpSession session) {
+        if(session.getAttribute("environmentalManager") ==  null) {
+            model.addAttribute("userLogin", new UserLogin() {});
+            session.setAttribute("nextUrl", "/service/delete/" + nameOfService);
+            return "redirect:/inicio/login";
+        }
         serviceDao.deleteService(nameOfService);
         return "redirect:../list";
     }
